# openapi3.yml
openapi: 3.0.0
info:
  title: API для сканера
  version: 1.0.1
  description: |
    API для управления сервисом

servers:
  - url: http://localhost:3000 # Замените на реальный URL вашего сервера
    description: Локальный сервер разработки

paths:
  # GET запросы
  /folder-structure:
    get:
      summary: Возвращает структуру папок в виде JSON
      description: Возвращает структуру папок (и файлов) в указанном каталоге в виде JSON.
      parameters:
        - in: query
          name: path
          schema:
            type: string
          required: true
          description: Абсолютный или относительный путь к папке.
      responses:
        '200':
          description: Успешный запрос. Возвращает структуру папок в виде JSON.
          content:
            application/json:
              schema:
                type: object
                
                properties:
                  status:
                    type: string
                  message:
                    type: string
                    example: "folder-structure:"
                    description: Сообщение, подтверждающее успешное получение структуры папок.
                  data:
                    type: object
                    description: Объект, содержащий структуру папок.
                    properties:
                      folders:
                        type: array
                        items:
                          type: object
                          properties:
                            name:
                              type: string
                              description: Имя папки.
                              example: "sh скрипты"
                            type:
                              type: string
                              description: Тип папки.
                              example: "folder-"
                        description: Массив объектов, представляющих папки.
                required:
                  - status
                  - message
                  - data
              example:
                {
                  "status": "folder-structure: completed",
                  "message": "",
                  "data": {
                    "folders": [
                      {
                        "name": "sh скрипты",
                        "type": "folder-"
                      },
                      {
                        "name": "7zip",
                        "type": "folder-"
                      },
                      {
                        "name": "UFW сетевой экран",
                        "type": "folder-"
                      },
                      {
                        "name": "скаченые",
                        "type": "folder-"
                      },
                      {
                        "name": "утилиты",
                        "type": "folder-"
                      },
                      {
                        "name": "чат_gpt",
                        "type": "folder-"
                      }
                    ]
                  }
                }
        '500': # Ошибка на сервере (например, не удалось создать папку из-за проблем с файловой системой)
          $ref: '#/components/responses/InternalServerError'

  /open-document:
    get:
      summary: Открыть document root и прочитьать ссылки
      description: Если папка содержит document папку внутри то читаем содержание в ней файлов и выводим
      parameters:
        - in: query
          name: path
          schema:
            type: string
          required: true
          description: Абсолютный или относительный путь к новой папке.
      responses:
        '201': # Обычно для создания ресурса используют 201 Created
          description: Возвращает массив строк с именами файлов.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
              example:
                - "строка 1"
                - "строка 2"
                - "и т.д."
        '500': # Ошибка на сервере (например, не удалось создать папку из-за проблем с файловой системой)
          $ref: '#/components/responses/InternalServerError'

  /open-themes:
    get:
      summary: Открыть document/themes и читает файлы in1.html и тд
      description: Открыть document/themes и читает файлы in1.html и тд
      parameters:
        - in: query
          name: path
          schema:
            type: string
          required: true
          description: Абсолютный или относительный путь к новой папке.
      responses:
        '201': # Обычно для создания ресурса используют 201 Created
          description: Возвращает массив строк с именами файлов.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
              example:
                - "строка 1"
                - "строка 2"
                - "и т.д."
        '500': # Ошибка на сервере (например, не удалось создать папку из-за проблем с файловой системой)
          $ref: '#/components/responses/InternalServerError'

  /scan:
    get:
      summary: Сканируем документ
      description: Сканируем документ и возвращаем результат обработки.
      responses:
        '200':
          description: Документ успешно отсканирован и обработан. Возвращает объект с сообщением об успехе и массивом обработанных строк.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
                    example: "Данные обработаны успешно"
                    description: Сообщение об успехе операции.
                  receivedData:
                    type: array
                    items:
                      type: string
                      description: Строка, полученная после обработки сканированного документа.
                    description: Массив строк, полученных в результате обработки сканированного документа.
                required:
                  - status
                  - message
                  - data
              example:
                {
                  "status": "scan: completed",
                  "message": "Данные обработаны успешно",
                  "data": [
                    "мне кажится алгоритм немного запутаный давай попробуеме по другому у нас есто некие переменые start и end когда мы начинаем перебирать строку мы имеем start=0 считаем 70 символов и после этого идем далее и ищем пробел. Если пробе найден фиксируем это как конечную точку и копируем строку от позиции start до этой точки в масив. В следующем просчоте start станет точкой которуюя мы нашли а новая опять высчитается . Таким образом мы собирем в масив из строк которые будут больше 70 . Потом останется только собрать все элементы масива поставив знак переноса. Мне кажится этот код легче отлаживать в случии каких то не учтных ошибок. Мы будем иметь масив выписаных кусков который можим посмотреть",
                    "добавь если нужна закругка модулей, и сделай документацию к функции в виде коментария"
                  ]
                }

  /view-page:
    get:
      summary: Посмотреть содержаие страницы
      description: Посмотреть содержаие страницы по ID
      parameters:
        - in: query
          name: id
          schema:
            type: integer
            format: int64 #  (или int32, в зависимости от диапазона ваших ID)
            description: Идентификатор li (должен быть числом).
            example: 1
          required: true
      responses:
        '200':
          description: Станица готова для просмотра
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "Folder open"
                example:
                  {
                    "status": "success",
                    "message": "pageview"
                  }
        '500': # Ошибка на сервере (например, не удалось создать папку из-за проблем с файловой системой)
          $ref: '#/components/responses/InternalServerError'

  # POST запросы
  /create-folder:
    post:
      summary: Создает новую папку
      description: Создает новую папку по указанному пути.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                path:
                  type: string
                  description: Абсолютный или относительный путь к новой папке.
                  example: "/путь/к/новой/папке"
              required:
                - path
      responses:
        '201':
          description: Папка успешно создана.
          content:
            application/json:
              schema:
                type: object
                properties:               
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "Folder created successfully"
                example:
                  {
                    "status": "success",
                    "message": "Folder created successfully"
                  }
        '500':
          $ref: '#/components/responses/InternalServerError'

  /create-page:
    post: # Изменено с GET на POST
      summary: Создаем страницы и темы из доукумента
      description: создает папку в выбранном пути, копируя по шаблону.  Использует указанный шаблон (indexTheme) при создании страницы.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                path:
                  type: string
                  description: Абсолютный или относительный путь к новой папке.
                  example: "/путь/к/новой/папке"
                indexTheme:
                  type: integer  # Изменено на integer, если это всегда целое число
                  description: Индекс темы, используемой для создания страницы.  Соответствует определенному шаблону.
                  example: 1 # Пример значения индекса темы
              required:
                - path
                - indexTheme  # Добавлено требование к indexTheme
      responses:
        '201': # Обычно для создания ресурса используют 201 Created
          description: Папка успешно создана.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "Folder created successfully"
                example:
                  {
                    "status": "success",
                    "message": "Folder created successfully"
                  }
        '500': # Ошибка на сервере (например, не удалось создать папку из-за проблем с файловой системой)
          $ref: '#/components/responses/InternalServerError'

  /create-topic:
    post:
      summary: Создает новую тему
      description: Создает новую тему на страницу
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  format: int64 # Или int32
                  description: Идентификатор новой страницы.
                  example: 123
                path:
                  type: string
                  description: Абсолютный или относительный путь к новой странице.
                  example: "/путь/к/новой/странице"
              required:
                - id
                - path
      responses:
        '201':
          description: Тема успешно создана.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "Folder created successfully"
                example:
                  {
                    "status": "success",
                    "message": "Folder open"
                  }
        '500':
          $ref: '#/components/responses/InternalServerError'

  /delete-list:
    post:
      summary: Удаляем тему по ID
      description: Удаляем тему по ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  format: int64
                  description: Идентификатор папки (должен быть числом).
                  example: 123
              required:
                - id
      responses:
        '200':
          description: Папка успешно открыта.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "Folder open"
                example:
                  {
                    "status": "success",
                    "message": "Folder open"
                  }
        '500': # Ошибка на сервере (например, не удалось создать папку из-за проблем с файловой системой)
          $ref: '#/components/responses/InternalServerError'

  /delete-select:
    post:
      summary: Удалит все что выделено галочками
      description: Создает новую тему на страницу
      requestBody:
        required: true
        content:
          application/json:
            schema:
             type: array  # Изменяем тип на array
             items:  # Описываем элементы массива
              type: string
      responses:
        '201':
          description: Удалили все выделенное галочками успешно.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "File modified"
                example:
                  {
                    "status": "success",
                    "message": "File modified"
                  }
        '500':
          $ref: '#/components/responses/InternalServerError'

  /leave-selected:
    post:
      summary: Оставим выделеное остальное удаляем
      description: Оставим выделеное остальное удаляем
      requestBody:
        required: true
        content:
          application/json:
            schema:
             type: array  # Изменяем тип на array
             items:  # Описываем элементы массива
              type: string
      responses:
        '201':
          description: Удалили все выделенное галочками успешно.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "File modified"
                example:
                  {
                    "status": "success",
                    "message": "File modified"
                  }
        '500':
          $ref: '#/components/responses/InternalServerError'

  /look-page:
    post:
      summary: На выделеные чексбуксы формируем страницы для просмотра
      description: На выделеные чексбуксы формируем страницы для просмотра
      requestBody:
        required: true
        content:
          application/json:
            schema:
             type: array  # Изменяем тип на array
             items:  # Описываем элементы массива
              type: string
      responses:
        '201':
          description: Страница сформирована
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "File modified"
                example:
                  {
                    "status": "success",
                    "message": "File create"
                  }
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Сообщение об ошибке.
          example: "Something went wrong"
      required:
        - error

  responses:  # Многократно используемые ответы
    BadRequest:
      description: Некорректный запрос.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            {
              "error": "Invalid request"
            }
    Forbidden:
      description: Доступ запрещен.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            {
              "error": "Access denied"
            }
    InternalServerError:
      description: Ошибка на сервере.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            {
              "error": "Internal server error"
            }